+++ NEWS.md
@@ -3,6 +3,7 @@
 * Initialised package.
 * Added a `NEWS.md` file to track changes to the package.
 * Added a function to work out beddays and admissions per month.
+* Added a function to extend population estimates.